Create a prototype tracker
Addresses
---------
Learning of go, building an MVP

Why
---
I chose Go because it can build single-file executables for all OSs.
The tracker should be an easy-to-install tool and not have any expectations
about the environment it runs in beyond the OS itself.

The data that's collected by the tracker are:
* timestamp at client location
* User
* context: Reading vs Problem solving vs browsing, etc
* action: click, keypress, etc

Future data items are:
* app: the application that has current focus.
* data: a json file that can be any data that the client chooses to send.
* client: an identifier string for the client
* sessionID: a unique identifier supplied by the client.
* custom: name=value pair

The data will be in a csv format, with the named columns in fixed order,
followed by the custom ones.